Rename HTTP Void Adapter
